streaming/requirements-streaming.txt
# Live Streaming Dashboard - Streaming Dependencies
# Add these to your existing requirements.txt or install separately

# Spark Streaming
pyspark>=3.4.0

# Kafka
kafka-python>=2.0.2
confluent-kafka>=2.3.0  # Alternative Kafka client

# Snowflake Spark Connector (install via spark-submit --packages or download JAR)
# net.snowflake:spark-snowflake_2.12:2.12.0-spark_3.4
# net.snowflake:snowflake-jdbc:3.14.4

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0

# MLB Data Source
pybaseball>=2.2.7

# Environment
python-dotenv>=1.0.0

# FastAPI for Simulation Control API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# HTTP Client (for dashboard to call API)
requests>=2.31.0

# Snowflake Connector (for simulation producer)
snowflake-connector-python>=3.5.0
cryptography>=41.0.0

# Optional: Flink (if using PyFlink)
# apache-flink>=1.18.0
